Adds a plugin facility for XPath functions
