+++ _delb/xpath/functions.py
@@ -15,84 +15,44 @@
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, Callable, Union
+from typing import TYPE_CHECKING, Any
+
+from _delb.plugins import plugin_manager
 
 if TYPE_CHECKING:
     from _delb.xpath.ast import EvaluationContext
 
 
-xpath_functions = {}
-
-
-# TODO move to PluginManager
-def register_xpath_function(arg: Union[Callable, str]) -> Callable:
-    """
-    Custom functions can be defined as shown in the following example. The first
-    argument to a function is always the evaluation context which holds the properties
-    ``node``, ``position``, ``size`` and ``namespaces``; followed by the expression's
-    arguments.
-
-    .. testcode::
-
-        from delb import register_xpath_function, Document
-        from _delb.xpath.ast import EvaluationContext
-
-
-        @register_xpath_function("is-last")
-        def is_last(context: EvaluationContext) -> bool:
-            return context.position == context.size
-
-        @register_xpath_function
-        def lowercase(_, string: str) -> str:
-            return string.lower()
-
-
-        document = Document("<root><node/><node foo='BAR'/></root>")
-        assert document.xpath("/*[is-last() and lowercase(@foo)='bar']").size == 1
-    """
-    if isinstance(arg, str):
-
-        def wrapper(func):
-            xpath_functions[arg] = func
-            return func
-
-        return wrapper
-
-    if callable(arg):
-        xpath_functions[arg.__name__] = arg
-        return arg
-
-
-@register_xpath_function
+@plugin_manager.register_xpath_function
 def concat(*strings: str) -> str:
     return "".join(strings)
 
 
-@register_xpath_function
+@plugin_manager.register_xpath_function
 def contains(_, string: str, substring: str) -> bool:
     return substring in string
 
 
-@register_xpath_function
+@plugin_manager.register_xpath_function
 def boolean(_, value: Any) -> bool:
     return bool(value)
 
 
-@register_xpath_function
+@plugin_manager.register_xpath_function
 def last(context: EvaluationContext) -> int:
     return context.size
 
 
-@register_xpath_function("not")
+@plugin_manager.register_xpath_function("not")
 def _not(_, value: Any) -> bool:
     return not value
 
 
-@register_xpath_function
+@plugin_manager.register_xpath_function
 def position(context: EvaluationContext) -> int:
     return context.position
 
 
-@register_xpath_function("starts-with")
+@plugin_manager.register_xpath_function("starts-with")
 def starts_with(_, string: str, prefix: str) -> bool:
     return string.startswith(prefix)
